RavenDB-22229 - fix legacy index replication
