+++ src/Raven.Server/Documents/Handlers/Admin/Processors/Indexes/AdminIndexHandlerProcessorForStaticPut.cs
@@ -1,12 +1,6 @@
-﻿using System.Threading.Tasks;
-using JetBrains.Annotations;
-using Raven.Client.Documents.Smuggler;
+﻿using JetBrains.Annotations;
 using Raven.Server.Documents.Indexes;
 using Raven.Server.ServerWide.Context;
-using Raven.Server.Smuggler.Documents;
-using Raven.Server.Smuggler.Documents.Data;
-using Raven.Server.Smuggler.Migration;
-using Sparrow.Json;
 
 namespace Raven.Server.Documents.Handlers.Admin.Processors.Indexes;
 
@@ -17,21 +11,4 @@ internal sealed class AdminIndexHandlerProcessorForStaticPut : AbstractAdminInde
     }
 
     protected override AbstractIndexCreateController GetIndexCreateProcessor() => RequestHandler.Database.IndexStore.Create;
-
-    protected override async ValueTask HandleLegacyIndexesAsync()
-    {
-        using (ContextPool.AllocateOperationContext(out JsonOperationContext jsonOperationContext))
-        await using (var stream = new ArrayStream(RequestHandler.RequestBodyStream(), nameof(DatabaseItemType.Indexes)))
-        using (var source = new StreamSource(stream, jsonOperationContext, RequestHandler.DatabaseName))
-        {
-            var destination = RequestHandler.Database.Smuggler.CreateDestination();
-            var options = new DatabaseSmugglerOptionsServerSide
-            {
-                OperateOnTypes = DatabaseItemType.Indexes
-            };
-
-            var smuggler = RequestHandler.Database.Smuggler.Create(source, destination, jsonOperationContext, options);
-            await smuggler.ExecuteAsync();
-        }
-    }
 }
