+++ src/Raven.Server/Documents/Sharding/Handlers/Admin/Processors/Indexes/ShardedAdminIndexHandlerProcessorForJavaScriptPut.cs
@@ -1,4 +1,6 @@
-﻿using JetBrains.Annotations;
+﻿using System.Threading.Tasks;
+using System;
+using JetBrains.Annotations;
 using Raven.Server.Config;
 using Raven.Server.Documents.Handlers.Admin.Processors.Indexes;
 using Raven.Server.Documents.Indexes;
@@ -15,4 +17,9 @@ internal sealed class ShardedAdminIndexHandlerProcessorForJavaScriptPut : Abstra
     protected override AbstractIndexCreateController GetIndexCreateProcessor() => RequestHandler.DatabaseContext.Indexes.Create;
 
     protected override RavenConfiguration GetDatabaseConfiguration() => RequestHandler.DatabaseContext.Configuration;
+
+    protected override ValueTask HandleIndexesFromLegacyReplicationAsync()
+    {
+        throw new NotSupportedException("Legacy replication of indexes isn't supported in a sharded environment");
+    }
 }
